update: move interface analyzers to a folder
